fix: split files to satisfy max-lines ratchet
Move Codex hook trust helpers and the agentPresenceChanged client test out of oversized files so PR CI's max-lines check stays green.
