+++ test/unit/terminal_host_client_test.dart
@@ -278,36 +278,6 @@ void main() {
     expect(event.payload, <String, Object?>{'projectId': 'project-1'});
   });
 
-  test('forwards agentPresenceChanged on runtimeEvents', () async {
-    final tempDir = await Directory.systemTemp.createTemp(
-      'alera-host-client-agent-presence-',
-    );
-    addTearDown(() async {
-      if (await tempDir.exists()) {
-        await tempDir.delete(recursive: true);
-      }
-    });
-    final server = await _TerminalHostTestServer.start();
-    addTearDown(server.dispose);
-    final client = SocketTerminalHostClient(
-      launcher: _FakeTerminalHostLauncher(server: server),
-      applicationSupportDirectory: () async => tempDir,
-    );
-    addTearDown(client.dispose);
-
-    await client.ensureStarted(config: TerminalHostConfig.defaults);
-    final runtimeEvent = client.runtimeEvents.first;
-
-    server.send(<String, Object?>{
-      'event': 'agentPresenceChanged',
-      'payload': <String, Object?>{},
-    });
-
-    final event = await runtimeEvent;
-    expect(event.name, 'agentPresenceChanged');
-    expect(event.payload, <String, Object?>{});
-  });
-
   test(
     'configure updates connected hosts but does not start idle ones',
     () async {
